moved selection paint to paintSelectionBackground()
I made these changes while experimenting with the idea of
showing a larger selection rectangle than the entity itself,
so that the selected state would be visible even when the
interior is normally the same color or is obscured by an
image.  However, in the end I decided that was too ugly,
so all that's left is a little refactoring.