draw selection boxes right before their controller
This undoes a previous change I made where I drew all of the
selection boxes first.  That was motivated by the failed
experiment to draw selection boxes bigger than their entities--
I did not want those overlapping the adjacent entities.
Now that that's gone, it does not matter.

But drawing selection boxes first is really bad when you
have one entity above another: as soon as it is selected, it
seems to go behind the other.  So this restores the property
that the selection box Z-position is the same as the entity's.