made an attempt to add UI tests with Abbot; mostly failure
There is one simple test and it works.  But my attempts to
make more interesting tests ran into bugs and limitations
in Abbot.  I was able to fix some of them in Abbot itself,
but I ran out of steam with that approach.

My current thinking is to abandon Abbot and try another UI
test system or just build my own event record and playback
facility.