started on UI mockup project; added Window shape
As a first step toward making ded a UI mockup tool, I added
a "Window" shape option to Entity.  Its main feature at the
moment is the ability to automatically move and resize
contained elements using a "resize center".  That feature is
important because when doing dialog editing, it is common to
need to make a window a little bigger or smaller, and modifying
all the contained elements individually is tedious.