rewrote PNG writing code to fix error handling and symlinks
The original code just called the convenience wrapper
"ImageIO.write", but that does a bad job of reporting
file access errors, and does not work with symlinks
because it removes the target file first.

The new code uses the API one layer below that one,
which solves those file access problems.  It also is
the first step toward being able to write PNG comments.