add HTML anchor names to entities, and insert-ded-image-map script; bump to v12
This adds the ability to specify an HTML anchor name for each
entity in a diagram, and a Python script that knows how to
read those names and create an HTML image map for the
diagram.

Adding the anchor name bumps the file format to 12.

There is now a right-click on entities to set the anchor
names all at once.

Control labels are now specified *without* the colon so I can
use the label name in a help bubble caption.