add my own bitmap font rendering; but old is not yet removed
In this state, all the diagram text is rendered *twice*, once using
the old system and then again with the new, so that any single-pixel
differences will be visible.

Now I need to remove the old system and do some refactoring.

Also, this checkin includes LiteralLineMetrics.java, which I ended up
not using, so plan to remove in the next checkin.