insert-ded-image-map: fix error handling in parse_ded_file
The error handling in parse_ded_file was wrong.  The only way the
handler for KeyError could be reached is if the "entities" attribute
was missing from the ded file, but in that case, the 'entities' local
variable would not be set.

I also made the 'die' call in handle_entity more compact; I see no
reason to spread it over four lines instead of one.